refactor : messageRes 수정
