--- presentation/src/main/kotlin/com/silvertown/android/dailyphrase/presentation/ui/ActionType.kt
@@ -1,5 +1,0 @@
-package com.silvertown.android.dailyphrase.presentation.ui
-
-enum class ActionType {
-    LIKE, BOOKMARK, SHARE, NONE
-}
